Color4,Point2,Point3,Rect: Add explicit keyword on copy constructor
* Add explicit keyword on Rect, Point2, Point3 template classes

* Remove explicit copy assignment operators; these should never be
necessary.

* Renderer: Fix compile err created due to explicit keyword set on
Point2 — a sign that the explicit keyword is working as intended! :-)

+++ include/nomlib/graphics/Renderer.hpp
@@ -80,7 +80,7 @@ class Renderer
     const IntRect viewport ( void ) const;
 
     /// Obtain the current scale factor
-    const Point2i scale_factor ( void ) const;
+    const Point2f scale_factor ( void ) const;
 
     /// Obtain the blending mode used for drawing
     const SDL_BlendMode blend_mode ( void ) const;
